feat: adding jsdoc and adding a common function interface, and some minor chores
