+++ packages/huddle01-interfaces/README.md
@@ -1,1 +1,9 @@
-# videomeet-interface
+# Huddle01 Interfaces
+
+`@dapp-sdk/huddle01-interfaces` contains all the abstract classes for various messaging methods.
+
+Mainly this package has a common interface and two special classes.
+
+- `CommonInterface`: These are the methods that are common for both the special classes.
+- `AudioSpacesInterface`: This class includes all the methods required to build an audio spaces using Huddle01 services.
+- `VideoMeetInterface`: This class includes all the methods required to build a video meeting app utilising Huddle01 services.
