+++ packages/huddle01-interfaces/src/index.ts
@@ -1,4 +1,5 @@
 import AudioSpacesInterface from './AudioSpacesInterface';
 import VideoMeetInterface from './VideoMeetInterface';
+import { TRoomControls } from './CommonFunctionInterface';
 
-export { AudioSpacesInterface, VideoMeetInterface };
+export { AudioSpacesInterface, VideoMeetInterface, type TRoomControls };
