+++ packages/messaging-huddle01/README.md
@@ -1,1 +1,8 @@
-# spaces-huddle01
+# Messaging Huddle01
+
+Messaging Huddle01 is a package to help you build video/audio apps with ease.
+
+The package `@dapp-sdk/messagin-huddle01` exports two clients for you to build your app on.
+
+- `VideoMeetClient`: This client helps user access the methods required to create and interact in a video meeting app.
+- `AudioSpacesClient`: This client helps user access the methods required to create and interface in an audio only meeting.
