+++ packages/messaging-huddle01/src/index.ts
@@ -1,4 +1,7 @@
-import AudioSpacesClient from './AudioSpacesClient';
-import VideoMeetClient from './VideoMeetClient';
+import AudioSpacesClient, { ICreateSpacesResponse } from './AudioSpacesClient';
+
+import VideoMeetClient, { ICreateMeetResponse } from './VideoMeetClient';
 
 export { AudioSpacesClient, VideoMeetClient };
+
+export type { ICreateMeetResponse, ICreateSpacesResponse };
